requirements.txt
streamlit
diffusers
transformers
torch
pillow
scipy
accelerate
